pyproject.toml
[project]
name = "tg-say-my-post"
version = "0.1.0"
description = "Telegram TTS bot with synthesis stub"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0.0",
  "pydub>=0.25.1",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "pyproject.toml"]



